+++ src/entities/partyroom-info/index.ts
@@ -1,0 +1,5 @@
+export {
+  default as PartyroomMutationForm,
+  type PartyroomMutationSubmitHandler,
+} from './ui/form.component';
+export type { Model as PartyroomMutationFormModel } from './model/form.model';
